[Bronze V] Title: 이상한 기호, Time: 80 ms, Memory: 11532 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,9 +7,9 @@ public class Main{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
         
-        System.out.println((int)Math.pow(a,2) - (int)Math.pow(b,2));
+        System.out.println((long)Math.pow(a,2) - (long)Math.pow(b,2));
     }
 }
